requirements.txt
torchvision==0.11.1
pandas==1.3.4
matplotlib==3.4.3
timm